tests: fixtures: Move MCTP response logic to Endpoint
We're currently handling the MCTP messaging in the MCTPSocket object -
arguably closer to the source, but it's really the endpoints that
conceptually handle the requests and generate responses.

Instead, move the MCTP message handling to the Endpoint object.

We'll use this in a future test where we need to modify the Control
Protocol responder behaviour.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>